pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "git-agent"
version = "0.1.0"
dependencies = [
    "fastmcp",
    "gitpython",
]

[tool.setuptools.packages.find]
where = ["src"]